refactor!: remove NumberFieldLabel
fix new-york and default path for components
