+++ apps/www/src/lib/registry/default/example/NumberFieldPercentage.vue
@@ -5,8 +5,8 @@ import {
   NumberFieldDecrement,
   NumberFieldIncrement,
   NumberFieldInput,
-  NumberFieldLabel,
-} from '@/lib/registry/new-york/ui/number-field'
+} from '@/lib/registry/default/ui/number-field'
+import { Label } from '@/lib/registry/default/ui/label'
 </script>
 
 <template>
@@ -17,7 +17,7 @@ import {
       style: 'percent',
     }"
   >
-    <NumberFieldLabel>Percent</NumberFieldLabel>
+    <Label>Percent</Label>
     <NumberFieldContent>
       <NumberFieldDecrement />
       <NumberFieldInput />
