+++ apps/www/src/lib/registry/default/ui/number-field/index.ts
@@ -1,6 +1,5 @@
 export { default as NumberField } from './NumberField.vue'
 export { default as NumberFieldInput } from './NumberFieldInput.vue'
-export { default as NumberFieldLabel } from './NumberFieldLabel.vue'
 export { default as NumberFieldIncrement } from './NumberFieldIncrement.vue'
 export { default as NumberFieldDecrement } from './NumberFieldDecrement.vue'
 export { default as NumberFieldContent } from './NumberFieldContent.vue'
